Remove `$libdir` prefix, test Postgres 17 and 18
The `$libdir` prefix is not necessary, and prevents Postgres from
finding shared modules in other directories in `dynamic_library_path`.
So remove it. (The `module_pathname` directive is superfluous, since
pair includes no shared modules, but it's often used as an example
distribution, so leave it).

Also add Postgres 17 & 18 to the test matrix, update the copyright date,
add a discussion of `make instal prefix=` to the README, and switch to
fenced code blocks on the Markdown files.